lessterm: use Terminal=true
Terminal=true actually means "please use gnome-terminal".

There is unfortunately no way around this, it is hard coded in xdg-open
or whatever library is used to open .desktop files.

The only (dirty) solution, is to not have gnome-terminal installed, and
to create a symbolic link on the system PATH that points to the actual
terminal emulator that you want to use.

In my case:

  sudo ln -sf /usr/bin/footclient /usr/local/bin/gnome-terminal

Signed-off-by: Robin Jarry <robin@jarry.cc>